feat: OP Stack integration test - cleanups and refactors
