+++ e2e/setup.go
@@ -1,4 +1,4 @@
-package e2e_test
+package e2e
 
 import (
 	"context"
